fix(模型汇率): 手录一条价,这一页就会 500——一个 SQL 字面量没转型
`admin_model_prices` 那条四分支 UNION 里,手录价那一支写的是 `'', 1.0, e.probe_ms`。
PostgreSQL 把裸字面量 `1.0` 定为 **numeric**,而 Rust 那一位声明的是 `f64` ——
sqlx 解码当场失败,整页 500。

`endpoint_model_price` 现在是 0 行,所以这一支返回不了任何东西,页面看起来是好的。
**只要有人手录一条价(界面上就有那个按钮)就炸。** 和 plan-health 那次一模一样的形状:
有数据才炸,空表上永远是绿的。

这条是上一个 commit 修好扫描器之后**当场扫出来的**——它此前落在
`src[:src.find("\n#[cfg(test)]")]` 那道切片之外,从来没被检查过。
去掉 `::float8` 就会红,变异验过。

+++ server/src/relay_rates.rs
@@ -625,7 +625,7 @@ pub async fn admin_model_prices(
     let manual: Vec<PriceRow> = sqlx::query_as(
         "SELECT e.id, p.model_id, e.base_url, m.label, e.label, \
                 p.input_per_mtok, p.output_per_mtok, p.cached_per_mtok, \
-                '', 1.0, e.probe_ms, e.probe_ok \
+                '', 1.0::float8, e.probe_ms, e.probe_ok \
          FROM endpoint_model_price p \
          JOIN route_endpoints e ON e.id = p.endpoint_id \
          JOIN models m ON m.id = e.route_id \
